backend/requirements.txt
Django==4.1.4
djangorestframework==3.13.1
python-dotenv==0.19.2
